+++ tests/django/test_async_view.py
@@ -94,7 +94,20 @@ async def test_fails_when_not_sending_query():
     with pytest.raises(SuspiciousOperation) as e:
         await AsyncGraphQLView.as_view(schema=schema)(request)
 
-        assert e.value.args == ("No GraphQL query found in the request",)
+    assert e.value.args == ("No GraphQL query found in the request",)
+
+
+async def test_fails_when_request_body_has_invalid_json():
+    factory = RequestFactory()
+
+    request = factory.post(
+        "/graphql/", "definitely-not-json-string", content_type="application/json"
+    )
+
+    with pytest.raises(SuspiciousOperation) as e:
+        await AsyncGraphQLView.as_view(schema=schema, graphiql=False)(request)
+
+    assert e.value.args == ("Unable to parse request body as JSON",)
 
 
 @pytest.mark.django_db